fix(dashboard): extract getAuthToken, harden clipboard + test cleanup
- Move getAuthToken() to shared utils/auth.ts (was duplicated in App.tsx and ConsolePage.tsx)
- Add .catch() to clipboard write + clear timeout on unmount
- Restore global stubs in afterEach/afterAll in ConsolePage tests
